# src/docs/openapi_Voter.yaml
openapi: 3.0.1
info:
  title: API de Votación
  contact:
    name: Francisco Robledo Venegas
    url: https://github.com/Riddikuluz/UTEM_API_REST
    email: francisco.robledov@utem.cl
  version: 1.0.0
paths:
  /votacion/resultados:
    get:
      summary: Obtener resultados de votación
      security:
        - aouthToken: []
      responses:
        '200':
          description: Respuesta exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '403':
          description: Datos del cliente no encontrados
    /votacion/ramos:
      get:
        summary: Obtener ramos de votación
        security:
          - aouthToken: []
        responses:
          '200':
            description: Respuesta exitosa
            content:
              application/json:
                schema:
                  type: array
                  items:
                    type: string
        '403':
          description: Datos del cliente no encontrados
    /votacion/{seccion_curso}/resultados:
      get:
        summary: Obtener resultados de votación para una sección de curso específica
        security:
          - aouthToken: []
        parameters:
          - name: seccion_curso
            in: path
            description: Número de sección del curso
            required: true
            schema:
              type: string
        responses:
          '200':
            description: Respuesta exitosa
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    votaciones:
                      type: array
                      items:
                        type: string
                    promedio_seccion:
                      type: number
          '403':
            description: Datos del cliente no encontrados
          '404':
            description: No hay registros
    /votacion/votar:
      post:
        summary: Registrar voto
        security:
          - aouthToken: []
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  fecha:
                    type: string
                    format: date
                  valoracion:
                    type: string
                  seccion_curso:
                    type: string
        responses:
          '200':
            description: Votación realizada
          '201':
            description: Ya votó por este ramo, espere al próximo día
          '404':
            description: Datos del cliente no encontrados
          '500':
            description: Error del servidor
security:
  - aouthToken: []
components:
  securitySchemes:
    aouthToken:
      type: apiKey
      name: Authorization
      in: header